requirements.txt
py-clob-client
python-dotenv
requests
pandas
numpy
